requirements.txt
fastapi
uvicorn
opencv-python
pillow
torch
transformers
